Split k8s resources into separate files
